Remove unnecessary casts

+++ src/code/code_801067F0.c
@@ -8,6 +8,6 @@ f32 func_801067F0(f32 arg0, f32 arg1) {
     if (arg1 == 0.0f) {
         return 0.0f;
     }
-    sp4 = (s32)(arg0 / arg1);
-    return arg0 - ((f32)sp4 * arg1);
+    sp4 = arg0 / arg1;
+    return arg0 - (sp4 * arg1);
 }
